gcf stdout request tests passing
